\name{elmer-package}
\alias{elmer-package}
\alias{elmer}
\docType{package}
\title{
Interface to ElmerParam.
}
\description{
Uses ElmerParam to define a function of a bunch of parameters, using
external computation programs such as ElmerSolver.
}
\details{
\tabular{ll}{
Package: \tab elmerparam\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-08-17\cr
License: \tab GPL\cr
}
This package provides the function elmer_param().
}
\author{
Erik Edelmann <erik.edelmann@csc.fi>
}
\references{
The ElmerParam Manual, ElmerParam/doc
}
\keyword{Elmer}
\seealso{
}
\examples{
}
